\name{JSDEsim1}
\alias{JSDEsim1}
\docType{data}
\title{
Simulated Jump Diffusion
}
\description{
Simulated Jump Diffusion
}
\usage{data("JSDEsim1")}
\format{
  A data frame with 1001 observations on the following 3 variables.
  \describe{
    \item{\code{X_t}}{a numeric vector}
    \item{\code{Y_t}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}

\keyword{datasets}
