\name{RcppArmadillo-Functions}
\alias{junkfunction2}
\title{A Junk Funktion For Build Purposes}
\description{
  This function was created as a filler in order for the package to build correctly.
}
\usage{
junkfunction2()
}
\value{
  \code{junkfunction2()} does nothing useful.
 
}
\details{
  This function was created as a filler in order for the package to build correctly. 
}
\references{
  See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\author{Etienne A.D. Pienaar}

